packages/orchestrator/pyproject.toml
[project]
name = "scanner-orchestrator"
version = "0.1.0"
description = "API FastAPI, workers Celery, base de données"
requires-python = ">=3.12"
dependencies = [
    "scanner-shared",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "celery[redis]>=5.3",
    "minio>=7.2",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "python-multipart>=0.0.26",
]

[project.optional-dependencies]
# Ces dépendances sont lourdes (plusieurs centaines de Mo) et ne sont
# utiles que dans les workers pipeline, pas dans le conteneur API.
# Installer avec : uv sync --all-packages --extra pipeline
pipeline = [
    "open3d>=0.18",
    "pymeshlab>=2023.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scanner_orchestrator"]

[tool.uv.sources]
scanner-shared = { workspace = true }
